[Продвинутый трек][Ревью] 2-й семинар

+++ advanced/src/main/java/academy/backend/market_pulse/model/Bond.java
@@ -1,20 +1,21 @@
 package academy.backend.market_pulse.model;
 
-import java.math.BigDecimal;
-
+/**
+ * Облигация.
+ */
 public class Bond extends Instrument {
 
-    private final BigDecimal couponRate;
+    private final double couponRate;
     private final int maturityYear;
 
     public Bond(String ticker, String name, Currency currency,
-                BigDecimal couponRate, int maturityYear) {
+                double couponRate, int maturityYear) {
         super(ticker, name, currency);
         this.couponRate = couponRate;
         this.maturityYear = maturityYear;
     }
 
-    public BigDecimal getCouponRate() {
+    public double getCouponRate() {
         return couponRate;
     }
